Moves shaders and classes into ModelExperimental
