+++ Source/Scene/ModelExperimental/ModelExperimentalUtility.js
@@ -1,10 +1,10 @@
-import BoundingSphere from "../Core/BoundingSphere.js";
-import Cartesian3 from "../Core/Cartesian3.js";
-import defined from "../Core/defined.js";
-import Matrix4 from "../Core/Matrix4.js";
-import Quaternion from "../Core/Quaternion.js";
-import RuntimeError from "../Core/RuntimeError.js";
-import Axis from "./Axis.js";
+import BoundingSphere from "../../Core/BoundingSphere.js";
+import Cartesian3 from "../../Core/Cartesian3.js";
+import defined from "../../Core/defined.js";
+import Matrix4 from "../../Core/Matrix4.js";
+import Quaternion from "../../Core/Quaternion.js";
+import RuntimeError from "../../Core/RuntimeError.js";
+import Axis from "../Axis.js";
 /**
  * @private
  */
